Created `updated.py` for all updating functions. Updated `.env.example`.
