+++ .env.example
@@ -2,10 +2,10 @@ ENV=DEVELOPMENT
 
 # Logging
 LOG_LEVEL=DEBUG
-LOG_RSS_ENTRIES=false
+LOG_RSS_ENTRIES=true
 LOG_TIPS=true
 
 # Development
-WATCHER_DIR=./
-DOWNLOADS_DIR=./downloads
+WATCHER_DIR=./src
+DOWNLOADS_DIR=./src/downloads
 INTERVAL_SEC=60
